+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Added
+
+- new HTMLNode
+- new ChartNode
+
+### Changed
+
+- elements refactoring
+- bumped sdk version
+- yFiles prop name in SimpleEdge class
+
+### Removed
+
+- unnecessary primitives
+
 ## [0.5.0]
 
 ### Added